Split globe chamber into responsive dynamic views
